Introduce NotFound
finds now returns a NotFound object, which is a thin wrapper around
Int.  This allows easy determination between "found" and "should be
inserted at this position".  This simplifies a lot of logic.

inserts now returns the position where the object was actually
inserted.  This *can* be different from the given position if
duplicates occur in the list: duplicates will always be added
at the end.

Adapt documentation and tests accordingly